+++ clients/fides-js/src/components/Overlay.tsx
@@ -1,10 +1,6 @@
 import { h, FunctionComponent, VNode } from "preact";
 import { useEffect, useState, useCallback, useMemo } from "preact/hooks";
-import {
-  FidesOptions,
-  PrivacyExperience,
-  ServingComponent,
-} from "../lib/consent-types";
+import { FidesOptions, PrivacyExperience } from "../lib/consent-types";
 
 import { debugLog, hasActionNeededNotices } from "../lib/consent-utils";
 
@@ -13,11 +9,7 @@ import { useA11yDialog } from "../lib/a11y-dialog";
 import ConsentModal from "./ConsentModal";
 import { useHasMounted } from "../lib/hooks";
 import { dispatchFidesEvent } from "../lib/events";
-import {
-  FidesCookie,
-  getOrMakeFidesCookie,
-  isNewFidesCookie,
-} from "../lib/cookie";
+import { FidesCookie } from "../lib/cookie";
 
 interface RenderBannerProps {
   isOpen: boolean;
@@ -39,16 +31,6 @@ interface Props {
   renderModalContent: (props: RenderModalContent) => VNode;
 }
 
-/**
- * Overlay doesn't always have the most up to date cookie. For the most part,
- * we prefer the cookie on the DOM, except when that cookie doesn't exist yet,
- * in which case we want the cookie object that was passed into this component.
- */
-export const getLatestCookie = (cookieFromParam: FidesCookie) => {
-  const latestCookie = getOrMakeFidesCookie();
-  return isNewFidesCookie(latestCookie) ? cookieFromParam : latestCookie;
-};
-
 const Overlay: FunctionComponent<Props> = ({
   experience,
   options,
@@ -63,11 +45,7 @@ const Overlay: FunctionComponent<Props> = ({
   const [bannerIsOpen, setBannerIsOpen] = useState(false);
 
   const dispatchCloseEvent = useCallback(() => {
-    dispatchFidesEvent(
-      "FidesModalClosed",
-      getLatestCookie(cookie),
-      options.debug
-    );
+    dispatchFidesEvent("FidesModalClosed", cookie, options.debug);
   }, [cookie, options.debug]);
 
   const { instance, attributes } = useA11yDialog({