+++ clients/fides-js/src/components/tcf/TcfOverlay.tsx
@@ -35,7 +35,10 @@ import {
   ServingComponent,
 } from "../../lib/consent-types";
 import { generateTcString } from "../../lib/tcf";
-import { FidesCookie } from "../../lib/cookie";
+import {
+  FidesCookie,
+  transformTcfPreferencesToCookieKeys,
+} from "../../lib/cookie";
 import InitialLayer from "./InitialLayer";
 import TcfTabs from "./TcfTabs";
 import Button from "../Button";
@@ -181,12 +184,16 @@ const updateCookie = async (
   oldCookie: FidesCookie,
   tcf: TcfSavePreferences,
   experience: PrivacyExperience
-) => {
+): Promise<FidesCookie> => {
   const tcString = await generateTcString({
     tcStringPreferences: tcf,
     experience,
   });
-  return { ...oldCookie, tcString };
+  return {
+    ...oldCookie,
+    tc_string: tcString,
+    tcf_consent: transformTcfPreferencesToCookieKeys(tcf),
+  };
 };
 
 const TcfOverlay: FunctionComponent<OverlayProps> = ({